Check if the user is the owner of the restaurant menu

+++ final_project.py
@@ -4,8 +4,6 @@ import json
 import random
 import string
 
-import httplib2
-import requests
 from flask import (
     Flask, render_template, request, redirect, url_for, jsonify, flash,
     session as login_session, make_response
@@ -13,15 +11,17 @@ from flask import (
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from oauth2client.client import flow_from_clientsecrets, FlowExchangeError
+import httplib2
+import requests
 
 from database_setup import Base, User, Restaurant, MenuItem
 
-CLIENT_SECRET_FILE = 'client_secrets.json'
+client_secret_file = 'client_secrets.json'
 
-with open(CLIENT_SECRET_FILE, 'r') as f:
+with open(client_secret_file, 'r') as f:
     web_client_id = json.load(f)['web']['client_id']
 
-CLIENT_ID = web_client_id
+client_id = web_client_id
 
 app = Flask(__name__)
 
@@ -32,19 +32,26 @@ session = Session()
 
 
 @app.route('/')
-@app.route('/restaurants')
+@app.route('/restaurants/')
 def show_restaurants():
     restaurants = session.query(Restaurant).all()
     no_restaurants_msg = None
 
     if not restaurants:
-        no_restaurants_msg = "There are no restaurants listed yet. Add one!"
+        no_restaurants_msg = 'There are no restaurants listed yet. Add one!'
 
-    return render_template(
-        'restaurants.html',
-        restaurants=restaurants,
-        no_restaurants_msg=no_restaurants_msg
-    )
+    if 'username' not in login_session:
+        return render_template(
+            'public-restaurants.html',
+            restaurants=restaurants,
+            no_restaurants_msg=no_restaurants_msg,
+        )
+    else:
+        return render_template(
+            'restaurants.html',
+            restaurants=restaurants,
+            no_restaurants_msg=no_restaurants_msg,
+        )
 
 
 @app.route('/restaurant/new', methods=['GET', 'POST'])
@@ -60,13 +67,13 @@ def new_restaurant():
         if name:
             new_restaurant = Restaurant(
                 name=name,
-                user_id=user_id
+                user_id=user_id,
             )
 
             session.add(new_restaurant)
             session.commit()
 
-            flash("New Restaurant Created")
+            flash('New Restaurant Created')
 
             return redirect(url_for('show_restaurants'))
 
@@ -75,20 +82,27 @@ def new_restaurant():
 
 @app.route('/restaurant/<int:restaurant_id>/edit', methods=['GET', 'POST'])
 def edit_restaurant(restaurant_id):
+    edit_restaurant = session.query(Restaurant).filter_by(
+        id=restaurant_id,
+    ).one()
+
     # Check if user is signed in
     if 'username' not in login_session:
         return redirect('/login')
 
-    edit_restaurant = session.query(Restaurant).filter_by(
-        id=restaurant_id
-    ).one()
+    if edit_restaurant.user_id != login_session['user_id']:
+        return ("<script>function myFunction() {alert('You are not "
+                "authorized to edit this restaurant. "
+                "Please create your own restaurant in order "
+                "to edit.');}</script><body onload='myFunction()''>")
 
     if request.method == 'POST':
         name = request.form['name']
+
         if name:
             edit_restaurant.name = name
 
-            flash("Restaurant Successfully Edited")
+            flash('Restaurant Successfully Edited')
 
             return redirect(url_for('show_restaurants'))
 
@@ -97,57 +111,82 @@ def edit_restaurant(restaurant_id):
 
 @app.route('/restaurant/<int:restaurant_id>/delete', methods=['GET', 'POST'])
 def delete_restaurant(restaurant_id):
+    delete_restaurant = session.query(Restaurant).filter_by(
+        id=restaurant_id,
+    ).one()
+
     # Check if user is signed in
     if 'username' not in login_session:
         return redirect('/login')
 
-    delete_restaurant = session.query(Restaurant).filter_by(
-        id=restaurant_id
-    ).one()
+    if delete_restaurant.user_id != login_session['user_id']:
+        return ("<script>function myFunction() {alert('You are not "
+                "authorized to delete this restaurant. "
+                "Please create your own restaurant in order "
+                "to delete.');}</script><body onload='myFunction()''>")
 
     if request.method == 'POST':
         session.delete(delete_restaurant)
         session.commit()
 
-        flash("Restaurant Successfully Deleted")
+        flash('Restaurant Successfully Deleted')
 
         return redirect(url_for('show_restaurants'))
 
     return render_template(
         'delete-restaurant.html',
-        restaurant=delete_restaurant
+        restaurant=delete_restaurant,
     )
 
 
-@app.route('/restaurant/<int:restaurant_id>')
-@app.route('/restaurant/<int:restaurant_id>/menu')
+@app.route('/restaurant/<int:restaurant_id>/')
+@app.route('/restaurant/<int:restaurant_id>/menu/')
 def show_restaurant_menu(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
     creator = get_user_info(restaurant.user_id)
     menu_items = session.query(MenuItem).filter_by(
-        restaurant_id=restaurant_id
+        restaurant_id=restaurant_id,
     ).all()
     no_menu_items_msg = None
 
     if not menu_items:
-        no_menu_items_msg = "There are no items on this menu yet. Add some!"
+        no_menu_items_msg = 'There are no items on this menu yet. Add some!'
 
-    return render_template(
-        'menu.html',
-        restaurant=restaurant,
-        creator=creator,
-        items=menu_items,
-        no_items_msg=no_menu_items_msg
-    )
+    if ('username' not in login_session or
+            creator.id != login_session['user_id']):
+        return render_template(
+            'public-menu.html',
+            restaurant=restaurant,
+            creator=creator,
+            items=menu_items,
+            no_items_msg=no_menu_items_msg,
+        )
+    else:
+        return render_template(
+            'menu.html',
+            restaurant=restaurant,
+            creator=creator,
+            items=menu_items,
+            no_items_msg=no_menu_items_msg,
+        )
 
 
-@app.route('/restaurant/<int:restaurant_id>/menu/new', methods=['GET', 'POST'])
+@app.route(
+    '/restaurant/<int:restaurant_id>/menu/new',
+    methods=['GET', 'POST'],
+)
 def new_menu_item(restaurant_id):
+    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+
     # Check if user is signed in
     if 'username' not in login_session:
         return redirect('/login')
 
-    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    if login_session['user_id'] != restaurant.user_id:
+        return ("<script>function myFunction() {alert('You are not "
+                "authorized to add menu items to this restaurant. "
+                "Please create your own restaurant in order "
+                "to add items.');}</script><body onload='myFunction()''>")
 
     if request.method == 'POST':
         name = request.form['name']
@@ -160,17 +199,17 @@ def new_menu_item(restaurant_id):
             price=price,
             course=course,
             restaurant_id=restaurant_id,
-            user_id=restaurant.user_id
+            user_id=restaurant.user_id,
         )
 
         session.add(new_item)
         session.commit()
 
-        flash("Menu Item Created")
+        flash('Menu Item Created')
 
         return redirect(url_for(
             'show_restaurant_menu',
-            restaurant_id=restaurant_id
+            restaurant_id=restaurant_id,
         ))
 
     return render_template('new-menu-item.html', restaurant=restaurant)
@@ -178,17 +217,23 @@ def new_menu_item(restaurant_id):
 
 @app.route(
     '/restaurant/<int:restaurant_id>/menu/<int:menu_id>/edit',
-    methods=['GET', 'POST']
+    methods=['GET', 'POST'],
 )
 def edit_menu_item(restaurant_id, menu_id):
+    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    edit_item = session.query(MenuItem).filter_by(
+        id=menu_id,
+    ).one()
+
     # Check if user is signed in
     if 'username' not in login_session:
         return redirect('/login')
 
-    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
-    edit_item = session.query(MenuItem).filter_by(
-        id=menu_id
-    ).one()
+    if login_session['user_id'] != restaurant.user_id:
+        return ("<script>function myFunction() {alert('You are not "
+                "authorized to edit menu items to this restaurant. "
+                "Please create your own restaurant in order "
+                "to edit items.');}</script><body onload='myFunction()''>")
 
     if request.method == 'POST':
         name = request.form['name']
@@ -208,47 +253,53 @@ def edit_menu_item(restaurant_id, menu_id):
         if course:
             edit_item.course = course
 
-        flash("Menu Item Successfully Edited")
+        flash('Menu Item Successfully Edited')
 
         return redirect(url_for(
             'show_restaurant_menu',
-            restaurant_id=restaurant_id
+            restaurant_id=restaurant_id,
         ))
 
     return render_template(
         'edit-menu-item.html',
         restaurant=restaurant,
-        item=edit_item
+        item=edit_item,
     )
 
 
 @app.route(
     '/restaurant/<int:restaurant_id>/menu/<int:menu_id>/delete',
-    methods=['GET', 'POST']
+    methods=['GET', 'POST'],
 )
 def delete_menu_item(restaurant_id, menu_id):
+    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    delete_item = session.query(MenuItem).filter_by(id=menu_id).one()
+
     # Check if user is signed in
     if 'username' not in login_session:
         return redirect('/login')
 
-    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
-    delete_item = session.query(MenuItem).filter_by(id=menu_id).one()
+    if login_session['user_id'] != restaurant.user_id:
+        return ("<script>function myFunction() {alert('You are not "
+                "authorized to delete menu items to this restaurant. "
+                "Please create your own restaurant in order "
+                "to delete items.');}</script><body onload='myFunction()''>")
 
     if request.method == 'POST':
         session.delete(delete_item)
         session.commit()
 
-        flash("Menu Item Successfully Deleted")
+        flash('Menu Item Successfully Deleted')
 
         return redirect(url_for(
             'show_restaurant_menu',
-            restaurant_id=restaurant_id
+            restaurant_id=restaurant_id,
         ))
 
     return render_template(
         'delete-menu-item.html',
         restaurant=restaurant,
-        item=delete_item
+        item=delete_item,
     )
 
 
@@ -260,7 +311,7 @@ def show_login():
     ) for x in range(32))
     login_session['state'] = state
 
-    return render_template('login.html', CLIENT_ID=CLIENT_ID, STATE=state)
+    return render_template('login.html', CLIENT_ID=client_id, STATE=state)
 
 
 # Connect with Google login
@@ -270,7 +321,7 @@ def gconnect():
     if request.args.get('state') != login_session['state']:
         response = make_response(
             json.dumps('Invalid state parameter.'),
-            401
+            401,
         )
         response.headers['Content-type'] = 'application/json'
 
@@ -281,13 +332,13 @@ def gconnect():
 
     try:
         # Upgrade authorization code into credentials object
-        oauth_flow = flow_from_clientsecrets(CLIENT_SECRET_FILE, scope='')
+        oauth_flow = flow_from_clientsecrets(client_secret_file, scope='')
         oauth_flow.redirect_uri = 'postmessage'
         credentials = oauth_flow.step2_exchange(code)
     except FlowExchangeError:
         response = make_response(
             json.dumps('Failed to upgrade the authorization code.'),
-            401
+            401,
         )
         response.headers['Content-type'] = 'application/json'
 
@@ -295,7 +346,8 @@ def gconnect():
 
     # Check access token is valid
     access_token = credentials.access_token
-    auth_url = 'https://www.googleapis.com/oauth2/v1/tokeninfo?access_token={}'.format(access_token)
+    auth_url = ('https://www.googleapis.com/oauth2/v1/'
+                'tokeninfo?access_token={}'.format(access_token))
     h = httplib2.Http()
     result = json.loads(h.request(auth_url, 'GET')[1])
 
@@ -303,7 +355,7 @@ def gconnect():
     if result.get('error') is not None:
         response = make_response(
             json.dumps(result.get('error')),
-            500
+            500,
         )
         response.headers['Content-type'] = 'application/json'
 
@@ -315,17 +367,17 @@ def gconnect():
     if result['user_id'] != google_id:
         response = make_response(
             json.dumps("Token's user ID doesn't match given user ID."),
-            401
+            401,
         )
         response.headers['Content-type'] = 'application/json'
 
         return response
 
     # Verify access token is valid for this app
-    if result['issued_to'] != CLIENT_ID:
+    if result['issued_to'] != client_id:
         response = make_response(
             json.dumps("Token's client ID doesn't match app's ID."),
-            401
+            401,
         )
         print("Token's client ID doesn't match app's ID.")
         response.headers['Content-type'] = 'application/json'
@@ -339,7 +391,7 @@ def gconnect():
     if stored_access_token is not None and google_id == stored_google_id:
         response = make_response(
             json.dumps('Current user is already connected.'),
-            200
+            200,
         )
         response.headers['Content-type'] = 'application/json'
 
@@ -350,10 +402,10 @@ def gconnect():
     login_session['google_id'] = google_id
 
     # Get user info
-    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    userinfo_url = 'https://www.googleapis.com/oauth2/v1/userinfo'
     params = {
         'access_token': credentials.access_token,
-        'alt': 'json'
+        'alt': 'json',
     }
     user_request = requests.get(userinfo_url, params=params)
     user_data = user_request.json()
@@ -369,12 +421,15 @@ def gconnect():
 
     login_session['user_id'] = user_id
 
-    output = """<h1>Welcome, {}!</h1>
-        <img src="{}" style="width: 300px; height: 300px; border-radius: 50%;">
-    """.format(login_session['username'], login_session['picture'])
+    output = ('<h1>Welcome, {}!</h1><img src="{}" '
+              'style="width: 300px; height: 300px; '
+              'border-radius: 50%;">'.format(
+                login_session['username'],
+                login_session['picture'],
+              ))
 
-    flash("You are now logged in as {}".format(login_session['username']))
-    print("Done!")
+    flash('You are now logged in as {}'.format(login_session['username']))
+    print('Done!')
     return output
 
 
@@ -388,14 +443,15 @@ def gdisconnect():
     if access_token is None:
         response = make_response(
             json.dumps('Current user not connected.'),
-            401
+            401,
         )
         response.headers['Content-type'] = 'application/json'
 
         return response
 
     # Execute HTTP GET response to revoke current token
-    disconnect_url = 'https://accounts.google.com/o/oauth2/revoke?token={}'.format(access_token)
+    disconnect_url = ('https://accounts.google.com/o/oauth2/'
+                      'revoke?token={}'.format(access_token))
     h = httplib2.Http()
     result = h.request(disconnect_url, 'GET')[0]
 
@@ -409,7 +465,7 @@ def gdisconnect():
 
         response = make_response(
             json.dumps('Successfully disconnected.'),
-            200
+            200,
         )
         response.headers['Content-type'] = 'application/json'
 
@@ -418,7 +474,7 @@ def gdisconnect():
         # For any reason, the given token was invalid
         response = make_response(
             json.dumps('Failed to revoke token for given user.'),
-            400
+            400,
         )
         response.headers['Content-type'] = 'application/json'
 
@@ -429,14 +485,14 @@ def create_user(login_session):
     new_user = User(
         name=login_session['username'],
         email=login_session['email'],
-        picture=login_session['picture']
+        picture=login_session['picture'],
     )
 
     session.add(new_user)
     session.commit()
 
     user = session.query(User).filter_by(
-        email=login_session['email']
+        email=login_session['email'],
     ).one()
 
     return user.id
@@ -463,17 +519,17 @@ def restaurants_json():
     restaurants = session.query(Restaurant).all()
 
     return jsonify(
-        restaurants=[restaurant.serialize for restaurant in restaurants]
+        restaurants=[r.serialize for r in restaurants]
     )
 
 
 @app.route('/restaurant/<int:restaurant_id>/menu/JSON')
 def restaurant_menu_json(restaurant_id):
     menu_items = session.query(MenuItem).filter_by(
-        restaurant_id=restaurant_id
+        restaurant_id=restaurant_id,
     ).all()
 
-    return jsonify(items=[item.serialize for item in menu_items])
+    return jsonify(items=[i.serialize for i in menu_items])
 
 
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/JSON')
